third_party/WebKit/LayoutTests/http/tests/devtools/sources/outline-javascript-5.html

<html>
<head>
<script src="../../inspector/inspector-test.js"></script>
<script src="../../inspector/sources-test.js"></script>

<script>

async function test() {
  var test = SourcesTestRunner.testJavascriptOutline;

  TestRunner.markStep('testComputedProperty');
  await test('a.b[c] = function() { };');

  TestRunner.markStep('testComputedPropertyInTheMiddle');
  await test('a.b[c].d = function() { };');

  TestRunner.markStep('testComputedPropertyForExpression');
  await test('(a || b).c = function() { };');

  TestRunner.markStep('testPropertyStringLiteral');
  await test('var foo = { "bar": function() { }}');

  TestRunner.markStep('testPropertyNumberLiteral');
  await test('var foo = { 42: function() { }}');

  TestRunner.completeTest();
}

</script>

</head>

<body onload="runTest()">
<p>Verify javascript outline</p>
</body>
</html>
