third_party/WebKit/LayoutTests/http/tests/devtools/sources/outline-javascript-5-expected.txt

CONSOLE WARNING: line 3078: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Verify javascript outline


Running: testComputedProperty
Text:
    a.b[c] = function() { };
Outline:

Running: testComputedPropertyInTheMiddle
Text:
    a.b[c].d = function() { };
Outline:
    d():0:0

Running: testComputedPropertyForExpression
Text:
    (a || b).c = function() { };
Outline:
    c():0:0

Running: testPropertyStringLiteral
Text:
    var foo = { "bar": function() { }}
Outline:

Running: testPropertyNumberLiteral
Text:
    var foo = { 42: function() { }}
Outline:
